Ballot-169 Patch: Address False-Positives From Certain Webserver Configs
Updates "3.2.2.4.6 Agreed-Upon Change to Website" to prohibit including a
complete Request Token or Random Value in the URL of the change, so as to
prohibit implementations that could generate false-positive verifications
from certain webserver configurations.

This defines a new term "Required Website Content".

Credit for this text goes to Andrew Ayer.